Consolidate action/feat duplicate code into helpers
