+++ src/module/item/action/sheet.ts
@@ -1,8 +1,7 @@
 import { ActionItemPF2e } from "@item/action/document.ts";
 import { ItemSheetDataPF2e } from "@item/sheet/data-types.ts";
 import { ItemSheetPF2e } from "../sheet/base.ts";
-import { FrequencySource } from "@item/data/base.ts";
-import { htmlQuery } from "@util";
+import { addSheetFrequencyListeners } from "./helpers.ts";
 
 export class ActionSheetPF2e extends ItemSheetPF2e<ActionItemPF2e> {
     override async getData(options?: Partial<DocumentSheetOptions>): Promise<ActionSheetData> {
@@ -24,15 +23,7 @@ export class ActionSheetPF2e extends ItemSheetPF2e<ActionItemPF2e> {
     override activateListeners($html: JQuery<HTMLElement>): void {
         super.activateListeners($html);
         const html = $html[0];
-
-        htmlQuery(html, "a[data-action=frequency-add]")?.addEventListener("click", () => {
-            const frequency: FrequencySource = { max: 1, per: "day" };
-            this.item.update({ system: { frequency } });
-        });
-
-        htmlQuery(html, "a[data-action=frequency-delete]")?.addEventListener("click", () => {
-            this.item.update({ "system.-=frequency": null });
-        });
+        addSheetFrequencyListeners(this.item, html);
     }
 
     protected override _getSubmitData(updateData?: Record<string, unknown>): Record<string, unknown> {
