+++ src/module/item/feat/sheet.ts
@@ -1,9 +1,9 @@
+import { addSheetFrequencyListeners } from "@item/action/helpers.ts";
 import { FeatPF2e } from "@item/feat/document.ts";
 import { ItemSheetDataPF2e, ItemSheetPF2e } from "@item/sheet/index.ts";
 import { htmlQuery, tagify } from "@util";
 import Tagify from "@yaireo/tagify";
 import { featCanHaveKeyOptions } from "./helpers.ts";
-import { FrequencySource } from "@item/data/base.ts";
 
 class FeatSheetPF2e extends ItemSheetPF2e<FeatPF2e> {
     override get validTraits(): Record<string, string> {
@@ -35,6 +35,7 @@ class FeatSheetPF2e extends ItemSheetPF2e<FeatPF2e> {
     override activateListeners($html: JQuery<HTMLElement>): void {
         super.activateListeners($html);
         const html = $html[0];
+        addSheetFrequencyListeners(this.item, html);
 
         const prerequisites = htmlQuery<HTMLInputElement>(html, 'input[name="system.prerequisites.value"]');
         if (prerequisites) {
@@ -42,15 +43,6 @@ class FeatSheetPF2e extends ItemSheetPF2e<FeatPF2e> {
                 editTags: 1,
             });
         }
-
-        htmlQuery(html, "a[data-action=frequency-add]")?.addEventListener("click", () => {
-            const frequency: FrequencySource = { max: 1, per: "day" };
-            this.item.update({ system: { frequency } });
-        });
-
-        htmlQuery(html, "a[data-action=frequency-delete]")?.addEventListener("click", () => {
-            this.item.update({ "system.-=frequency": null });
-        });
 
         const keyOptionsInput = htmlQuery<HTMLInputElement>(html, 'input[name="system.subfeatures.keyOptions"]');
         tagify(keyOptionsInput, { whitelist: CONFIG.PF2E.abilities, maxTags: 3 });
